Move to a function override approach
Currently we are using variables to override values.

Variables aren't very flexible. This is making the codebase hard to read
to read and maintain.

This is a proposal commit to move to a function override based approach.

This approach would allow users to better configure their prompts and
plugins without having to fork and change the source code.