Inline lots of environment variables
The biggest change is going from := to :-
The downside is we dont have these nice list of environment variables
at the top of each function.

The upside is we are no longer making global assignments

The main difference between := and :- is that :- just substitutes
the word on the right hand side inline, while := assigns the variable
on the left hand side to the value on the right.

    echo ${A:=something} # assign A to 'something', then expand $A
    echo ${A:-something} # expand to 'something'

One reason this is hard to understand is the zsh manpages describe
:= and :- in an inverted manner - one description uses negatives
and exceptions, while the other uses positives. I'm not sure if its
more accurate but at least we aren't polluting variables as much.

This rabbit hole happened when I came across `setopt warn_global_assign`