Allow customization of terminal title
Add GEOMETRY_TITLE and GEOMETRY_CMDTITLE display locations.
Geometry was already controlling the terminal title. This change uses the geometry function plumbing to display information there.
The original explicit title-controlling code displayed the name and arguments of the currently-executing command. This change introduces a new geometry_cmd function to display this information from within the GEOMETRY_CMDTITLE configuration.

The riskiest change in this code is the hand-rolled deansi function. The colored output of geometry functions displays poorly in the terminal title, so non-text information needed to be removed.